Moved OSCALLoader to components and separated Loader form into its own OSCALLoaderForm component
