+++ src/components/OSCALProfile.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, fireEvent, act } from "@testing-library/react";
-import { OSCALProfileLoader } from "../OSCALLoader";
+import { OSCALProfileLoader } from "./OSCALLoader";
 import OSCALProfile from "./OSCALProfile";
 import { metadataTestData, testOSCALMetadata } from "./OSCALMetadata.test";
 import testOSCALBackMatter, {
